Reduce memory usage (#72)
* Minor fixes to JSON to PB conversion function

* Handle duplicate fields when creating the vid map from VCF headers

* Restructured code so that the next commit allows incremental
update of INFO fields without having to obtain reordered data from
all CallSets. This saves significant memory.
The primary change in this commit is to move the "sum" operation
to its own function which is called by a function that iterates over
the callsets. The incremental update will ensure that a single buffer
is used for reordering and the sum function is called immediately after
every callset reordering.

* Restructure the combined VCF generator to use the per sample remapping
and incremental combine (sum) operations. All the other code changes
are to achieve this goal.

* Completed incremental remapping and combine operation for INFO fields
to save memory. Bug fixes in this commit

* Added unit tests for duplicate field name checking in the VCF header

* Check for duplicates in FILTER field names also

* Minor edits to allow building with Protobuf versions 3.0.0-beta-1
and 3.0.x stable