Support cloud storage via htslib plugins better (#175)
GenomicsDB htslib plugin support is now initialized without a filename and it caches tiledb contexts for all the protocols it supports by maintaining a map between the uri prefix and tiledb context. The prefix for the uri is just the minimum part of the uri that points to the cloud storage account and container/bucket.