Integrate CodeQL with GitHub Actions for vulnerable analysis (#194)
Integrate CodeQL with GitHub Actions for static vulnerability scanning of C++ and Java sources. CodeQL does not seem to handle path/path-ignore filters for C++, so for now, we have to manually dismiss alerts for htslib, rapidjson, etc. We can look into building htslib/rapidjson/tiledb out of the source tree and see if that helps later.